Fix #2088: Clarify token colours and placement.
This updates the images of the arena to use colours similar to those
we've been using on the badges, and to clarify which tokens go where.
Modify the item about where the tokens go to explain the placement
scheme, and then use this scheme to update the colours in the diagrams.

Change-Id: I71b79b7d85e402a0dccdc7ac83a4e25b720cda41