Fix #2098: Tokens must be on the floor to be in a zone.
This seems to be easiest way to judge this, and most in-line with
the original intent and the current opions of those who have said
anything about it.

This also removes the redundant repetition of regulation 2.7 from
this rule, which ensures that the rules don't spill over onto a third
page (which would only contain a couple of lines).

Change-Id: Icaa8c7deae8e698369bc7434c7cedb955de5c6f1